requirements.txt
<<<<<<< HEAD
alembic==1X.13.1
=======
>>>>>>> f96a6c714a534eee9657deb83997051c69bed856
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.1.2
black==24.1.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dnspython==2.5.0
email-validator==2.1.0.post1
executing==1.2.0
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
Mako==1.3.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
mysqlclient==2.2.1
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.12.1
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.2
stripe==8.1.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.6
Werkzeug==3.0.1
WTForms==3.1.2
